refactor: move OverlayClassListProxy to the internal namespace
